C++ Int64/wxLongLong to represent SQLite3 INTEGER (#6965)
* C++ Int64 to represent SQLite3 INTEGER

* save return int64 instead

* save return int64 instead

* use int64 for accountdialog

* use int64 for key of Index_By_Id

* use int64 for entity id in assetdialog

* use int64 as id for Model.h

* use int64 as id for attachmentdialog

* typedef std::int64_t int64

* typedef wxLongLong_t int64

* typedef wxLongLong int64

* DB_xxx get & get_record with int64

* DB_xxx remove with int64

* wxLongLong.GetValue for html_template

* wxLongLong.GetValue for row_t

* json_writer.Int64

* use GetLastRowId() directly as we switch to int64/wxLongLong now

* use int64/wxLongLong in Model_Translink

* use int64/wxLongLong in Model_Translink and json_writer.Int64(xxx.GetValue())

* use int64/wxLongLong in Model_Account, Model_Att and mmframe

* use int64/wxLongLong in AssetsPanel and Option

* use int64/wxLongLong in AssetsPanel

* use int64/wxLongLong in AssetsPanel

* use int64/wxLongLong in attachmentdialog

* use int64/wxLongLong in Model_Account

* use int64/wxLongLong in currency, infotable and option

* use int64/wxLongLong

* use int64/wxLongLong

* use int64/wxLongLong

* use int64/wxLongLong

* use int64/wxLongLong

* Use int64/wxlonglong: refine AssetPanel->initVirtualListControl

* Use int64/wxlonglong: wxArrayInt64 for TAGS

* Use int64/wxlonglong: wxArrayInt64 for TAGS

* Use int64/wxlonglong: wxArrayInt64

* Use int64/wxlonglong: declare wxArray

* Use int64/wxlonglong: declare wxArray

* Use int64/wxlonglong: use std::vector<wxLongLong> as wxArrayInt64

* Use int64/wxlonglong: more int => int64

* Use int64/wxlonglong: more int => int64

* Use int64/wxlonglong: billsdepositspanel.initVirtualListControl int64

* Use int64/wxlonglong: xxx => xxx.GetValue() during calc

* Use int64/wxlonglong: xxx => xxx.GetValue() during calc

* Use int64/wxlonglong: xxx => xxx.GetValue() during calc

* Use int64/wxlonglong: int => int64 for Model_Budget

* Use int64/wxlonglong: int => int64 for budgeting & cate

* Use int64/wxlonglong: category relocate

* Use int64/wxlonglong: mmListBoxItem.getIndex

* Use int64/wxlonglong

* Use int64/wxlonglong

* Use int64/wxlonglong for filtertrans

* Use int64/wxlonglong for filtertrans

* Use int64/wxlonglong for filtertrans

* Use int64/wxlonglong for filtertrans

* Use int64/wxlonglong for fusedtransaction

* Use Model_Billsdeposits::split

* Use wxArrayInt64 and fix typo

* Use wxArrayInt64 and fix typo

* Use wxArrayInt64 and fix typo

* Use wxArrayInt64

* Use wxArrayInt64

* Use wxArrayInt64

* Use wxArrayInt64 for m_selected_categories_id

* Use wxArrayInt64 for m_selected_categories_id

* Use wxArrayInt64

* Use wxArrayInt64 for Model_Budgetsplittransaction

* Use wxArrayInt64

* Use wxArrayInt64 for GRM

* Use wxArrayInt64 for GRM

* Use wxArrayInt64 for currency

* Use wxArrayInt64 for currency

* Use wxArrayInt64 for currency

* Use wxArrayInt64 for Attachment

* Use wxArrayInt64 for transdialog and custom field

* Use wxArrayInt64 for mmchecking_list

* Use wxArrayInt64 for mmchecking_list, custom data and transaction update dlg

* Use wxArrayInt64 for QIF, csv and mmframe

* Use wxArrayInt64 for Model_Translink

* Use wxArrayInt64 for QIF

* Use wxArrayInt64 for QIF

* Use wxArrayInt64 for Wizard NewAccount and checkingpancel

* Use wxArrayInt64 for stockspanel

* Use wxArrayInt64 for home page

* Use wxArrayInt64 for home page and home page panel

* Use int64 for home page and let get support wxLongLong_t

* Use int64: fix typo

* Use int64: fix typo

* Use int64 and wxArrayInt64

* Use wxArrayInt64: const ref

* Use const auto ref

* revert themes submodule change

+++ src/assetdialog.cpp
@@ -360,8 +360,8 @@ void mmAssetDialog::OnOk(wxCommandEvent& /*event*/)
     m_asset->VALUECHANGERATE  = valueChangeRate;
     m_asset->ASSETTYPE        = asset_type;
 
-    int old_asset_id = m_asset->ASSETID;
-    int new_asset_id = Model_Asset::instance().save(m_asset);
+    int64 old_asset_id = m_asset->ASSETID;
+    int64 new_asset_id = Model_Asset::instance().save(m_asset);
 
     if (old_asset_id < 0)
     {
@@ -370,7 +370,7 @@ void mmAssetDialog::OnOk(wxCommandEvent& /*event*/)
     }
     if (m_transaction_panel->ValidCheckingAccountEntry())
     {
-        int checking_id = m_transaction_panel->SaveChecking();
+        int64 checking_id = m_transaction_panel->SaveChecking();
         if (checking_id < 0)
             return;
 
@@ -455,7 +455,7 @@ void mmAssetDialog::OnQuit(wxCloseEvent& /*event*/)
 void mmAssetDialog::OnAttachments(wxCommandEvent& /*event*/)
 {
     const wxString& RefType = Model_Attachment::reftype_desc(Model_Attachment::ASSET);
-    int RefId;
+    int64 RefId;
     
     if (!this->m_asset)
         RefId = 0;
