+++ src/model/Model_Taglink.cpp
@@ -51,7 +51,7 @@ Model_Taglink& Model_Taglink::instance()
     return Singleton<Model_Taglink>::instance();
 }
 
-Model_Taglink::Data* Model_Taglink::get(const wxString& refType, int refId, int tagId)
+Model_Taglink::Data* Model_Taglink::get(const wxString& refType, int64 refId, int64 tagId)
 {
     Data* link = this->get_one(REFTYPE(refType), REFID(refId), TAGID(tagId));
     if (link) return link;
@@ -61,9 +61,9 @@ Model_Taglink::Data* Model_Taglink::get(const wxString& refType, int refId, int 
     return link;
 }
 
-std::map<wxString, int> Model_Taglink::get(const wxString& refType, int refId)
+std::map<wxString, int64> Model_Taglink::get(const wxString& refType, int64 refId)
 {
-    std::map<wxString, int> tags;
+    std::map<wxString, int64> tags;
     for (const auto& link : instance().find(REFTYPE(refType), REFID(refId)))
         tags[Model_Tag::instance().get(link.TAGID)->TAGNAME] = link.TAGID;
 
@@ -71,7 +71,7 @@ std::map<wxString, int> Model_Taglink::get(const wxString& refType, int refId)
 }
 
 /* Delete all tags for a REFTYPE + REFID */
-void Model_Taglink::DeleteAllTags(const wxString& refType, int refID)
+void Model_Taglink::DeleteAllTags(const wxString& refType, int64 refID)
 {
     const auto& links = instance().find(REFTYPE(refType), REFID(refID));
     instance().Savepoint();
@@ -80,11 +80,11 @@ void Model_Taglink::DeleteAllTags(const wxString& refType, int refID)
     instance().ReleaseSavepoint();
 }
 
-int Model_Taglink::update(const Data_Set& rows, const wxString& refType, int refId)
+int Model_Taglink::update(const Data_Set& rows, const wxString& refType, int64 refId)
 {
     Model_Taglink::instance().Savepoint();
     bool updateTimestamp = false;
-    std::map<int, int> row_id_map;
+    std::map<int, int64> row_id_map;
 
     Data_Set links = instance().find(REFTYPE(refType), REFID(refId));
     if (links.size() != rows.size()) updateTimestamp = true;
@@ -134,11 +134,11 @@ int Model_Taglink::update(const Data_Set& rows, const wxString& refType, int ref
     return rows.size();
 }
 
-std::map<int, Model_Taglink::Data_Set> Model_Taglink::get_all(const wxString& refType)
+std::map<int64, Model_Taglink::Data_Set> Model_Taglink::get_all(const wxString& refType)
 {
     Data_Set taglinks = instance().find(REFTYPE(refType));
 
-    std::map<int, Data_Set> data;
+    std::map<int64, Data_Set> data;
     for (const auto& taglink : taglinks)
     {
         data[taglink.REFID].push_back(taglink);
